==> notebooks/1_creatingMarketPortfolio.ipynb <==
# Imports

import pandas as pd 
import numpy as np
import os
from pathlib import Path
from collections import Counter

dataPath = str(Path(os.getcwd()).parent.absolute())+"/data"

# Reading the data

stocksData = pd.read_csv(dataPath+'/raw/stocksData.csv', index_col=0) 

stocksData

# Creating a new dataframe for the market data

# Create the df
marketData = pd.DataFrame(index=stocksData.index, columns=["MKT|DAILY_RETURNS","MKT|MKT_CAP"])

# Calculating the Market cap

#create a list of all the stocks market cap column names
stocksMktCapColumns=[]
for column in stocksData:
    if column[-8:] == "|MKT_CAP":
        stocksMktCapColumns+=[column]

#sum all stocks market cap to have our market market cap
marketData['MKT|MKT_CAP'] = list(stocksData[stocksMktCapColumns].sum(axis=1))

# Calculating the daily returns

#creating the daily returns with the market cap changes
lastDayMktCap=marketData["MKT|MKT_CAP"]["2016-09-13"] #for day one it doesn't work so we set it to 0% (we'll delete the first row later)
for index, row in marketData.iterrows():
    marketData["MKT|DAILY_RETURNS"][index]=100*((marketData["MKT|MKT_CAP"][index]/lastDayMktCap)-1)
    lastDayMktCap=marketData["MKT|MKT_CAP"][index]


# Saving

marketData.to_csv(dataPath+'/interim/marketData.csv')
