==> notebooks/4b_dummyVariablesRegressions.ipynb <==
### Imports

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import statsmodels.api as sm
import shap
import sklearn
import timeit
import seaborn as sns
import os
from pathlib import Path
import warnings

dataPath = str(Path(os.getcwd()).parent.absolute())+"/data"
figuresPath = str(Path(os.getcwd()).parent.absolute())+"/reports/figures"

def significance(x):
    if x>0.1:return("")
    elif x>0.05:return("*")
    elif x>0.01:return("**")
    else:return("***")

sectors = pd.read_excel(dataPath+'/raw/sectors.xlsx', index_col=0, sheet_name="Sheet2") 

GMB_MKT_daily = pd.read_csv(dataPath+'/interim/marketGreenPortfolios.csv', index_col=0).loc[:,["MKT_GMB_BL|DAILY_RETURNS","MKT_GMB_RO|DAILY_RETURNS","MKT_GMB|DAILY_RETURNS"]]
# GMB_MKT_daily.head()
GMB_INDW_daily = pd.read_csv(dataPath+'/interim/industryWeightedGreenPortfolios.csv', index_col=0)[['IND_W_GREEN|DAILY_RETURNS','IND_W_BROWN|DAILY_RETURNS',"IND_W_GMB_BL|DAILY_RETURNS","IND_W_GMB_RO|DAILY_RETURNS","IND_W_GMB|DAILY_RETURNS"]]
#  GMB_INDW_daily.head()
MKT_daily = pd.read_csv(dataPath+'/interim/marketData.csv', index_col=0)["MKT|DAILY_RETURNS"]
# MKT_daily.head()
STOCKS_daily = pd.read_csv(dataPath+'/raw/stocksData.csv', index_col=0) 
#  STOCKS_daily.head()
FF_daily = pd.read_csv(dataPath+'/raw/ffData.csv', index_col=0)
# ff_ret.head()

G_INDEXES_daily = pd.read_csv(dataPath+'/raw/greenIndexes.csv', index_col=0) 
# G_INDEXES_daily.head()

IND_daily = pd.DataFrame()
for sector in sectors.index:
    IND_daily[str(sector)+"|DAILY_RETURNS"]=pd.read_csv(dataPath+"/interim/industries/industry"+str(sector)+'MarketData.csv', index_col=0)["MKT|DAILY_RETURNS"]
    IND_daily[str(sector)+"_BROWN_BL|DAILY_RETURNS"]=pd.read_csv(dataPath+"/interim/industries/industry"+str(sector)+'GreenPortfolios.csv', index_col=0)["MKT_BROWN_BL|DAILY_RETURNS"]
    IND_daily[str(sector)+"_BROWN_RO|DAILY_RETURNS"]=pd.read_csv(dataPath+"/interim/industries/industry"+str(sector)+'GreenPortfolios.csv', index_col=0)["MKT_BROWN_RO|DAILY_RETURNS"]
    IND_daily[str(sector)+"_BROWN|DAILY_RETURNS"] = (IND_daily[str(sector)+"_BROWN_RO|DAILY_RETURNS"]+IND_daily[str(sector)+"_BROWN_BL|DAILY_RETURNS"])/2
    IND_daily[str(sector)+"_GREEN_BL|DAILY_RETURNS"]=pd.read_csv(dataPath+"/interim/industries/industry"+str(sector)+'GreenPortfolios.csv', index_col=0)["MKT_GREEN_BL|DAILY_RETURNS"]
    IND_daily[str(sector)+"_GREEN_RO|DAILY_RETURNS"]=pd.read_csv(dataPath+"/interim/industries/industry"+str(sector)+'GreenPortfolios.csv', index_col=0)["MKT_GREEN_RO|DAILY_RETURNS"]
    IND_daily[str(sector)+"_GREEN|DAILY_RETURNS"] = (IND_daily[str(sector)+"_GREEN_RO|DAILY_RETURNS"]+IND_daily[str(sector)+"_GREEN_BL|DAILY_RETURNS"])/2



sectors = pd.read_excel(dataPath+'/raw/sectors.xlsx', index_col=0, sheet_name="Sheet2") 

G_INDEXES_daily

#creating dummy for covid
FF_daily['COV1_1Jan2021'] = FF_daily.loc[:, 'Mkt-RF']
FF_daily['COV2_1Jan2021'] = FF_daily.loc[:, 'Mkt-RF']

# FF_daily.loc[:'2020-11-03', 'COV1_1Jan2021'] = 0  #Biden election
# FF_daily.loc['2020-11-04':, 'COV1_1Jan2021'] = 1
# FF_daily.loc[:'2020-11-03', 'COV2_1Jan2021'] = 0

# FF_daily.loc[:'2021-01-01', 'COV1_1Jan2021'] = 0  #1st january 2021 (end of covid repercussions ?)
# FF_daily.loc['2021-01-02':, 'COV1_1Jan2021'] = 1
# FF_daily.loc[:'2021-01-04', 'COV2_1Jan2021'] = 0

# FF_daily.loc[:'2020-01-31', 'COV1_1Jan2021'] = 0  # COVID starts (WHO Issues Global Health Emergency)
# FF_daily.loc['2020-02-01':, 'COV1_1Jan2021'] = 1
# FF_daily.loc[:'2020-01-31', 'COV2_1Jan2021'] = 0

FF_daily.loc[:'2020-03-11', 'COV1_1Jan2021'] = 0  # COVID starts nb2 (WHO Declares COVID-19 a Pandemic)
FF_daily.loc['2020-03-12':, 'COV1_1Jan2021'] = 1
FF_daily.loc[:'2020-03-11', 'COV2_1Jan2021'] = 0

# FF_daily.loc[:, 'COV1_1Jan2021'] = 0  # No dummy variables
# FF_daily.loc[:, 'COV2_1Jan2021'] = 0

# FF_daily.head()
# FF_daily.tail()


merged=pd.concat([FF_daily,MKT_daily,IND_daily,GMB_MKT_daily,GMB_INDW_daily,G_INDEXES_daily],axis=1)
merged=merged.dropna()# drops the first row which contains NaN
# merged.head()

merged

portfolios = ['MKT|DAILY_RETURNS',
            '10|DAILY_RETURNS',
#             '10_BROWN_BL|DAILY_RETURNS',
#             '10_BROWN_RO|DAILY_RETURNS',
#             '10_BROWN|DAILY_RETURNS',
#             '10_GREEN_BL|DAILY_RETURNS',
#             '10_GREEN_RO|DAILY_RETURNS',
#             '10_GREEN|DAILY_RETURNS',
            '15|DAILY_RETURNS',
#             '15_BROWN_BL|DAILY_RETURNS',
#             '15_BROWN_RO|DAILY_RETURNS',
#             '15_BROWN|DAILY_RETURNS',
#             '15_GREEN_BL|DAILY_RETURNS',
#             '15_GREEN_RO|DAILY_RETURNS',
#             '15_GREEN|DAILY_RETURNS',
            '20|DAILY_RETURNS',
#             '20_BROWN_BL|DAILY_RETURNS',
#             '20_BROWN_RO|DAILY_RETURNS',
#             '20_BROWN|DAILY_RETURNS',
#             '20_GREEN_BL|DAILY_RETURNS',
#             '20_GREEN_RO|DAILY_RETURNS',
#             '20_GREEN|DAILY_RETURNS',
#             '25|DAILY_RETURNS',
#             '25_BROWN_BL|DAILY_RETURNS',
#             '25_BROWN_RO|DAILY_RETURNS',
#             '25_BROWN|DAILY_RETURNS',
#             '25_GREEN_BL|DAILY_RETURNS',
#             '25_GREEN_RO|DAILY_RETURNS',
#             '25_GREEN|DAILY_RETURNS',
#             '30|DAILY_RETURNS',
#             '30_BROWN_BL|DAILY_RETURNS',
#             '30_BROWN_RO|DAILY_RETURNS',
#             '30_BROWN|DAILY_RETURNS',
#             '30_GREEN_BL|DAILY_RETURNS',
#             '30_GREEN_RO|DAILY_RETURNS',
#             '30_GREEN|DAILY_RETURNS',
            '35|DAILY_RETURNS',
#             '35_BROWN_BL|DAILY_RETURNS',
#             '35_BROWN_RO|DAILY_RETURNS',
#             '35_BROWN|DAILY_RETURNS',
#             '35_GREEN_BL|DAILY_RETURNS',
#             '35_GREEN_RO|DAILY_RETURNS',
#             '35_GREEN|DAILY_RETURNS',
#             '40|DAILY_RETURNS',
#             '40_BROWN_BL|DAILY_RETURNS',
#             '40_BROWN_RO|DAILY_RETURNS',
#             '40_BROWN|DAILY_RETURNS',
#             '40_GREEN_BL|DAILY_RETURNS',
#             '40_GREEN_RO|DAILY_RETURNS',
#             '40_GREEN|DAILY_RETURNS',
#             '45|DAILY_RETURNS',
#             '45_BROWN_BL|DAILY_RETURNS',
#             '45_BROWN_RO|DAILY_RETURNS',
#             '45_BROWN|DAILY_RETURNS',
#             '45_GREEN_BL|DAILY_RETURNS',
#             '45_GREEN_RO|DAILY_RETURNS',
#             '45_GREEN|DAILY_RETURNS',
            '50|DAILY_RETURNS',
#             '50_BROWN_BL|DAILY_RETURNS',
#             '50_BROWN_RO|DAILY_RETURNS',
#             '50_BROWN|DAILY_RETURNS',
#             '50_GREEN_BL|DAILY_RETURNS',
#             '50_GREEN_RO|DAILY_RETURNS',
#             '50_GREEN|DAILY_RETURNS',
#             '55|DAILY_RETURNS',
#             '55_BROWN_BL|DAILY_RETURNS',
#             '55_BROWN_RO|DAILY_RETURNS',
#             '55_BROWN|DAILY_RETURNS',
#             '55_GREEN_BL|DAILY_RETURNS',
#             '55_GREEN_RO|DAILY_RETURNS',
#             '55_GREEN|DAILY_RETURNS',
#             '60|DAILY_RETURNS',
#             '60_BROWN_BL|DAILY_RETURNS',
#             '60_BROWN_RO|DAILY_RETURNS',
#             '60_BROWN|DAILY_RETURNS',
#             '60_GREEN_BL|DAILY_RETURNS',
#             '60_GREEN_RO|DAILY_RETURNS',
#             '60_GREEN|DAILY_RETURNS',
#             'MKT_GREEN_BL|DAILY_RETURNS',
#             'MKT_BROWN_BL|DAILY_RETURNS',
#             'MKT_GMB_BL|DAILY_RETURNS',
#             'MKT_GREEN_RO|DAILY_RETURNS',
#             'MKT_BROWN_RO|DAILY_RETURNS',
#             'MKT_GMB_RO|DAILY_RETURNS',
#             'MKT_BROWN|DAILY_RETURNS',
#             'MKT_GREEN|DAILY_RETURNS',
#             'MKT_GMB|DAILY_RETURNS',
#             'IND_W_GREEN_BL|DAILY_RETURNS',
#             'IND_W_BROWN_BL|DAILY_RETURNS',
#             'IND_W_GMB_BL|DAILY_RETURNS',
#             'IND_W_GREEN_RO|DAILY_RETURNS',
#             'IND_W_BROWN_RO|DAILY_RETURNS',
#             'IND_W_GMB_RO|DAILY_RETURNS',
#             'IND_W_BROWN|DAILY_RETURNS',
#             'IND_W_GREEN|DAILY_RETURNS',
#             'IND_W_GMB|DAILY_RETURNS',
#             'SPX|DAILY_RETURNS',
            'QGREEN|DAILY_RETURNS',
            'SXWESGU|DAILY_RETURNS',
            'RENIXX|DAILY_RETURNS',
            'CELS|DAILY_RETURNS',
            'ECO|DAILY_RETURNS',
            'SPGTCED|DAILY_RETURNS',
            'F3OILG|DAILY_RETURNS',
            'SXEP|DAILY_RETURNS',
            'DJ_OIL_AND_GAS|DAILY_RETURNS']




warnings.filterwarnings("ignore")
for portfolio in portfolios:
    X = merged[['Mkt-RF','COV1_1Jan2021','COV2_1Jan2021']]
    X1= sm.add_constant(X)   
    y = merged[portfolio]-merged['RF']
    model= sm.OLS(y,X1).fit()
    summaryMKT= pd.DataFrame(data=model.summary2().tables[1][["Coef.","P>|t|"]])
    summaryMKT["P>|t|"]=summaryMKT["P>|t|"].apply(significance)
#     summaryMKT["Significance"]=summaryMKT["P>|t|"].apply(significance)
    summaryMKT = summaryMKT.transpose()[["const","Mkt-RF","COV1_1Jan2021","COV2_1Jan2021"]]
    summaryMKT["Adj. R²"]=[model.rsquared_adj,""]
    summaryMKT = pd.concat([summaryMKT], keys=['MKT'], names=['Model'])


    X = merged[['Mkt-RF','SMB','HML','RMW','CMA','COV1_1Jan2021','COV2_1Jan2021']]
    X1= sm.add_constant(X)   
    y = merged[portfolio]-merged['RF']
    model= sm.OLS(y,X1).fit()
    summary5FF= pd.DataFrame(data=model.summary2().tables[1][["Coef.","P>|t|"]])
    summary5FF["P>|t|"]=summary5FF["P>|t|"].apply(significance)
#     summary5FF["Significance"]=summary5FF["P>|t|"].apply(significance)
    summary5FF = summary5FF.transpose()[["const","Mkt-RF","COV1_1Jan2021","COV2_1Jan2021"]]
    summary5FF["Adj. R²"]=[model.rsquared_adj,""]
    summary5FF = pd.concat([summary5FF], keys=['5FF'], names=['Model'])


    X = merged[['Mkt-RF','SMB','HML','RMW','CMA','IND_W_GMB|DAILY_RETURNS','COV1_1Jan2021','COV2_1Jan2021']]
    X1= sm.add_constant(X)   
    y = merged[portfolio]-merged['RF']
    model= sm.OLS(y,X1).fit()
    summary6FF = pd.DataFrame(data=model.summary2().tables[1][["Coef.","P>|t|"]])
    summary6FF["P>|t|"]=summary6FF["P>|t|"].apply(significance)
#     summary6FF["Significance"]=summary6FF["P>|t|"].apply(significance)
    summary6FF = summary6FF.transpose()[["const","Mkt-RF","COV1_1Jan2021","COV2_1Jan2021"]]
    summary6FF["Adj. R²"]=[model.rsquared_adj,""]
    summary6FF = pd.concat([summary6FF], keys=['6FF'], names=['Model'])

    summaryPortfolio = pd.concat([summaryMKT,summary5FF,summary6FF], axis=0)
    summaryPortfolio = pd.concat([summaryPortfolio], keys=[portfolio[:-14]], names=['Portfolio'])
    summaryPortfolio
    
    if portfolio == "MKT|DAILY_RETURNS":
        summary = summaryPortfolio.copy()
    else:
        summary = pd.concat([summary,summaryPortfolio], axis=0)

summary.rename(columns={'const': 'α', 'Mkt-RF': 'β','COV1_1Jan2021': 'δ', 'COV2_1Jan2021': 'γ'}, inplace=True)


pd.set_option('display.max_rows',300)
summary

summary.to_excel(dataPath+"/processed/regressionSummary.xlsx")
